# shared/api/orders/v1/orders.openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  version: 1.0.0
  description: API для управления заказами космических кораблей

servers:
  - url: http://localhost:8080

x-ogen:
  package: order
  target: ./shared/pkg/openapi/spaceship-orders/v1

paths:
  /api/v1/orders:
    post:
      summary: Создание заказа
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_uuid
                - part_uuids
              properties:
                user_uuid:
                  type: string
                  format: uuid
                part_uuids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                type: object
                required:
                  - order_uuid
                  - total_price
                properties:
                  order_uuid:
                    type: string
                    format: uuid
                  total_price:
                    type: number
                    format: double

  /api/v1/orders/{order_uuid}:
    get:
      summary: Получить заказ по UUID
      operationId: getOrderByUUID
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заказ найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Заказ не найден

  /api/v1/orders/{order_uuid}/pay:
    post:
      summary: Оплата заказа
      operationId: payOrder
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_method
              properties:
                payment_method:
                  $ref: '#/components/schemas/PaymentMethod'
      responses:
        '200':
          description: Успешная оплата
          content:
            application/json:
              schema:
                type: object
                required:
                  - transaction_uuid
                properties:
                  transaction_uuid:
                    type: string
                    format: uuid
        '404':
          description: Заказ не найден

  /api/v1/orders/{order_uuid}/cancel:
    post:
      summary: Отмена заказа
      operationId: cancelOrder
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Заказ успешно отменен
        '404':
          description: Заказ не найден
        '409':
          description: Конфликт, заказ уже оплачен

components:
  schemas:
    OrderStatus:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAID
        - CANCELLED

    PaymentMethod:
      type: string
      enum:
        - UNKNOWN
        - CARD
        - SBP
        - CREDIT_CARD
        - INVESTOR_MONEY

    Order:
      type: object
      required:
        - order_uuid
        - user_uuid
        - part_uuids
        - total_price
        - status
      properties:
        order_uuid:
          type: string
          format: uuid
        user_uuid:
          type: string
          format: uuid
        part_uuids:
          type: array
          items:
            type: string
            format: uuid
        total_price:
          type: number
          format: double
        transaction_uuid:
          type: string
          format: uuid
          nullable: true
        payment_method:
          allOf:
            - $ref: '#/components/schemas/PaymentMethod'
          nullable: true
        status:
          $ref: '#/components/schemas/OrderStatus'